requirements.txt
streamlit
google-generativeai
python-dotenv
SpeechRecognition
PyAudio
gTTS
# ... autres dépendances ...
pyttsx3